Update benchmark
The Rust code used in the benchmark had a few issues for reliable performance testing.

- It used `warp` as the HTTP server library. At OneSignal we use Axum, which is
  much faster.
- It did not use a `reqwest::Client`, meaning there was no HTTP connection
  re-use. This accounts for the bulk of the performance difference between the
  Rust and Deno code, which I assume is doing automatic connection reuse for
  `fetch()` calls.
- Moved benchmarks from Deno to `wrk`, a comprehensive HTTP testing tool

These taken together bring the performance trends much closer, with Rust
averaging 70.44ms response time and Deno 70.6ms. I believe the tight gap in
performance is largely due to the runtime being dominated by the performance of
the upstream endpoint. At the point this service was doing a lot of
serialization/deserialization or other CPU-bound work (I know the intention for
the gateway service would not have a lot of CPU work, but some will be
unavoidable), the gap will widen.